moves encryption to header
