+++ communication/diana-core/src/main/java/org/jnosql/diana/api/SymmetricSettingsEncryption.java
@@ -14,12 +14,7 @@
  *   Otavio Santana
  *
  */
-package org.jnosql.diana.api.encryption;
-
-import org.jnosql.diana.api.JNoSQLException;
-import org.jnosql.diana.api.Settings;
-import org.jnosql.diana.api.SettingsEncryption;
-import org.jnosql.diana.api.SettingsPriority;
+package org.jnosql.diana.api;
 
 import javax.crypto.Cipher;
 import javax.crypto.NoSuchPaddingException;
