--- communication/diana-core/src/main/java/org/jnosql/diana/api/encryption/package-info.java
@@ -1,21 +1,0 @@
-/*
- *
- *  Copyright (c) 2019 Otávio Santana and others
- *   All rights reserved. This program and the accompanying materials
- *   are made available under the terms of the Eclipse Public License v1.0
- *   and Apache License v2.0 which accompanies this distribution.
- *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
- *
- *   You may elect to redistribute this code under either of these licenses.
- *
- *   Contributors:
- *
- *   Otavio Santana
- *
- */
-
-/**
- * This package has implementations of {@link org.jnosql.diana.api.SettingsEncryption}
- */
-package org.jnosql.diana.api.encryption;
